MR fixes
