+++ src/utils/__mocks__/isBase64UrlImage.ts
@@ -1,3 +1,1 @@
-export const isBase64UrlImage = async (base64String: string) => {
-	return true;
-};
+export const isBase64UrlImage = async (base64String: string) => true;
