+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { BREAKPOINTS } from './breakpoints';
 export { isBase64UrlImage } from './isBase64UrlImage';
 export { generateDisabledControls, DISABLED_CONTROL } from './storybook';
+export { generateBase64FromImageFile } from './generateBase64FromImageFile';
